api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
aiofiles==24.1.0
email-validator==2.2.0
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.14.2
pyotp
qrcode[pil]
psutil==5.9.8
